+++ packages/adhere/changelog/CHANGELOG.md
@@ -8,6 +8,8 @@
   - 修改为返回的判断逻辑
 * **adhere-util**
   - color加入新方法
+* **adhere-ui-writingboard**
+  - 加入导出图片设置背景色
 
 # 1.0.21